fix(architecture): move diagnose_profile to config to resolve layer violation
